feat: add rename command for templates
