+++ .gitignore
@@ -8,3 +8,4 @@
 
 # Go
 bin
+testing
