+++ internal/generator/generator_test.go
@@ -1,0 +1,9 @@
+package generator
+
+import (
+	"testing"
+)
+
+func TestGenerateGitignore(t *testing.T) {
+	
+}
