+++ internal/structs/structs.go
@@ -15,7 +15,8 @@ type Template struct {
 
 func (c TemplateConfig) Default() config.Config {
 	return TemplateConfig{
-		Editor:    "nano",
+		Editor:    "code",
+		Base:      "",
 		Templates: []Template{},
 	}
 }
